Fix running under nightly: copy prebuilt `libtensorflow.so` to `OUT_DIR`
A change in cargo nightly (merged on February 7th 2017) changed how
dynamic libraries are loaded. This prevented the `tensorflow-sys` crate's
tests to run.

To fix, simply copy the `libtensorflow.so` pre-built library to `OUT_DIR`
and change `cargo:rustc-link-search=` path so cargo can find it when
running the tests.

See:
* issue: https://github.com/tensorflow/rust/issues/71
* cargo's pull request: https://github.com/rust-lang/cargo/pull/3651