Add a check in `test-all` to verify the TF versions in different location
The `run-valgrind` script fails silently if its Tensorflow version is not
the proper one. The reason is because the `export LD_LIBRARY_PATH` line
explicitly uses the TF version in the exported path. If it is wrong (for
example if the version defined in variable `tensorflow_version` near the
top of the file) then the first call to valgrind will fail.

All the output of the call to valgrind is sent to a log file, so the
failure is not reported. And since the first line of the script let it
abort on error (`set -e`), the script stop on the first call to valgrind.

This commit adds a check in the `test-all` script to make sure the version
values match where they appear. This should let the CI fail if the
TF versioning is wrong in the crates.