Allow Sessions to be used across threads
This implements Send and Sync for Session, which allows them to be shared across threads.  It also changes Session::run to take &self rather than &mut self, which allows Sessions to be used simulaneously in multiple threads without mutexes. This puts the responsibility on the caller to avoid race conditions.

Fixes #192